requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.2
pydantic-settings==2.6.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
celery==5.4.0
redis==5.2.0
httpx==0.28.0
mcp==1.1.2
python-multipart==0.0.17
websockets==14.1
requests==2.32.3
python-dotenv==1.0.1
aiofiles==24.1.0
psycopg[binary]==3.2.3
openai>=1.40.0  # Latest stable version
beautifulsoup4==4.12.3  # For HTML parsing in web crawler
weasyprint==66.0  # PDF generation
markdown==3.9  # Markdown to HTML conversion
numpy==1.26.4  # Numerical computing for evaluation metrics
scipy==1.11.4  # Statistical functions (pearsonr for CVSS correlation)